core: add input_stream::consume()
This method, intended for request parsers, receives a stream of temporary
buffers until it determines that parsing is done, at which point it can
return any unconsumed partial buffer back to the input stream.

It is intended as a replacement for read_until(), since it can parse in a
single pass.