build: add option to compile as PIE or SO.
This is the first step in a journy of many miles, to be able to run
Seastar applications (such as httpd or memcached) on top of OSv.

This patch adds a configure.py option "-so" to build the applications as
shared-objects instead of executables. It also adds the option "-pie" to
build position-independent executables instead.

These are needed to run a Seastar application on OSv, as OSv cannot run
a normal position-dependent executable. Note that currently, PIE won't
actually work (because of OSv bug #352 - unfortunately Seastar uses
thread_local in one place) - so only "-so" is useful to build an application
to run in OSv.

The resulting shared-object doesn't yet run on OSv (many bugs ahead, as
well as missing features), but it's a start.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>