virtio: make virt_to_phys a virtual function
In the existing code, virt_to_phys() was a fixed do-nothing function.
This is good for vhost, but not good enough in OSv where the to convert
virtual addresses to physical we need an actual calculation.

The solution in this patch, using a virtual function, is not optimal
and should probably be replaced with a template later.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>