xen: translate feature name
When the backend advertises "feature-rx-copy", the frontend should register for
"request-rx-copy". The local hypervisor seems to be forgiving about it, but the
one in AWS, it is not, and doubly so.

First, it doesn't recognize these as the same. And second, it refuses to
connect the backend if this feature is not advertised by the frontend.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>