build: thrift support
Support adding a thrift file as a source.  Since thrift generates multiple
output files, whose names cannot be trivially derived from the source file
name, we have to specify it as an object containing the source file name
and any additional information needed to derive the generated file names
(in this case, the generated thrift services).