thrift: drop use of scattered_message
The code attempts to use scattered_message::append_static() to gain zero
copy, claiming that _output protects the data's lifetime, but that's a
false claim - it will remain in the transmit queue waiting for an ACK
while we process (and respond to) the next request.

Fix by just writing the output naturally.  We may zero-copy support later,
though it is dubious it will help thrift, as the code is full of copies
and atomics.