test: test util/serialization.hh
Add a test (using the Boost unit-test framework) for the functions of
util/serealization.hh for serializing and de-serializing primitive types
like integers and strings.

The test checks that the round-trip is correct, i.e., taking an original
data, serializing it and de-serializing it back, gets the same data back.
Moreover, it compares the result of serialization to pre-computed expected
results (which were produced by a Java program), to confirm that our code
generates the same serialization as Java.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>