gossip: convert ApplicationState to C++
The gossiped application state is a list of key/version pairs. The keys
come from the application_state enum in this patch.

This is a trivial conversion of the Java enum. It is possible we may need
to add more features to this enum class (Java enums have all sorts of weird
capabilities beyond C++ enums), but with any luck, we won't...

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>