cql3: convert Cql.g to C++
Convert Cql.g to produce C++. To make things compile, convert the
useStatement grammar rule. Also add an ugly sed hack in configure.py to
support "#if 0" in the grammar file.

This patch also disables -Werror and enables -fpermissive for
antlr-generated code which attempts to convert a const pointer to a
non-const pointer without a cast:

  build/release/gen/cql3/CqlParser.cpp: In member function ‘void CqlParser::cfOrKsName(cql3::cf_name&, bool)’:
  build/release/gen/cql3/CqlParser.cpp:460:72: warning: invalid conversion from ‘const CommonTokenType* {aka const antlr3::CommonToken<antlr3::Traits<CqlLexer, CqlParser> >*}’ to ‘antlr3::Traits<CqlLexer, CqlParser>::CommonTokenType* {aka antlr3::CommonToken<antlr3::Traits<CqlLexer, CqlParser> >*}’ [-fpermissive]
               t =  this->matchToken(IDENT, &FOLLOW_IDENT_in_cfOrKsName153);

Signed-off-by: Pekka Enberg <penberg@cloudius-systems.com>