transport: de-ragel frame processing
Current ragel-based cql transport processing has a bug, in which it
assumes that frame headers will not cross a fragment boundary.  Rather than
fix it, drop ragel completely.  With a small complication, the frame header
size is known in advance, so we can just read_exactly() it.

The complication is that before the first frame is read, we do not know the
protocol version, which affects the frame header size.  Adjust for that by
reading the first byte, recording the version, and then splicing in the rest
of the frame header.