cql3: Use @context instead of @members in Cql.g
There is a difference in meaning of @members between Cpp and Java
targets, because of reasons which are not clear to me. In Cpp @members
section is not put inside parser/lexer class definition, so any
varaible definitions declared there become program-global variables
instead of instance-local. We should use @context instead.