build: fix annoying warning messages
It seems antlr has a bug: It takes code which looks like this in Cql.g:

	t=IDENT

and compiles it into the following in CqlParser.cpp:

	ImplTraits::CommonTokenType* t = NULL;
	...
	t =  this->matchToken(IDENT, &FOLLOW_IDENT_in_cident976);

But matchToken is declared as

	const CommonTokenType* matchToken( ANTLR_UINT32 ttype, BitsetListType* follow );

Obviously, assigning a const pointer into a non-const pointer generates a
warning, and on a clean compilation of Urchin we get dozens of these
annoying messages.

I'm not sure why Antlr has this bug and why nobody noticed it before or
who to report it to (I'm new to Antlr...), but it's easy to work around:
This patch adds to our build an ugly "sed" command which changes lines like

	ImplTraits::CommonTokenType* varname = NULL;

into
	const ImplTraits::CommonTokenType* varname = NULL;

and now Urchin compiles without annoying warning messages.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>